Leetcode/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-jwt-extended==4.3.1
pymongo==4.0.1
python-dotenv==0.19.1
docker==5.0.3
werkzeug==2.0.3
redis==4.5.5